# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Bitwarden client API contract (Cloudwarden)
  version: 0.1.0
  description: |-
    Wire contract for the Bitwarden client API as called by the official Bitwarden clients (web vault, browser extension, desktop, CLI). It exists so Cloudwarden can be built against a written contract instead of any server implementation.

    **Source of truth.** Every operation records the clients repository release tag it was read from in `x-source-version` and the file in `x-source-file`. The primary tag is `web-v2026.9.1` of github.com/bitwarden/clients. A few calls moved into the Bitwarden SDK (a separate repository) after 2025; those operations cite the older tag where the client still contained the call (`web-v2025.8.0` or `web-v2025.1.0`). Shapes that exist only inside the SDK (the `send_access` grant, enum values for policy types) are flagged in their descriptions.

    **Services and paths.** One base URL serves several services. Official clients derive: API `<base>/api`, identity `<base>/identity`, notifications `<base>/notifications`, icons `<base>/icons`, events `<base>/events`. Paths in this document include that prefix.

...

...

    **Errors.** API errors use `ErrorResponseModel` (`message`, `validationErrors`, `object: error`). Identity errors use OAuth fields plus `ErrorModel`. HTTP 429 is shown as a rate limit message. Successful empty responses may be 200 or 204; clients treat both as success.

...
  license:
    name: AGPL-3.0-only
...
  contact:
...
servers:
...
security:
- bearerAuth: []
...
paths:
  /api/accounts:
    delete:
      tags:
      - Accounts
      operationId: deleteAccount
      summary: Delete the account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretVerificationRequest'
      responses:
        '200':
          description: Success, empty body.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-source-version: web-v2026.9.1
      x-source-file: libs/common/src/auth/services/account-api.service.ts
  /api/accounts/api-key:
    post:
      tags:
      - Accounts
      operationId: getApiKey
      summary: Get the personal API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretVerificationRequest'
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-source-version: web-v2026.9.1
      x-source-file: libs/common/src/services/api.service.ts
...
  /api/accounts/email:
    post:
      tags:
      - Accounts
      operationId: changeEmail
      summary: Confirm an email change
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
      responses:
        '200':
          description: Success, empty body.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-source-version: web-v2026.9.1
      x-source-file: libs/common/src/auth/services/change-email/default-change-email.service.ts
  /api/accounts/email-token:
    post:
      tags:
      - Accounts
      operationId: requestEmailChangeToken
      summary: Request an email-change confirmation code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailTokenRequest'
      responses:
        '200':
          description: Success, empty body.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-source-version: web-v2026.9.1
      x-source-file: libs/common/src/auth/services/change-email/default-change-email.service.ts
  /api/accounts/kdf:
    post:
      tags:
      - Accounts
      operationId: changeKdf
      summary: Change KDF settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KdfRequest'
      responses:
        '200':
          description: Success, empty body.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-source-version: web-v2025.8.0
      x-source-file: libs/common/src/services/api.service.ts
  /api/accounts/key-management/rotate-user-account-keys:
    post:
      tags:
      - Accounts
      operationId: rotateUserAccountKeys
      summary: Rotate the user key and all wrapped data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RotateUserAccountKeysRequest'
      responses:
        '200':
          description: Success, empty body.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-source-version: web-v2026.9.1
      x-source-file: apps/web/src/app/key-management/key-rotation/user-key-rotation-api.service.ts
  /api/accounts/keys:
    post:
      tags:
      - Accounts
      operationId: setAccountKeys
      summary: Set the account key pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeysRequest'
      responses:
        '200':
          description: Success, empty body.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-source-version: web-v2026.9.1
      x-source-file: libs/common/src/services/api.service.ts
  /api/accounts/password:
    post:
      tags:
      - Accounts
      operationId: changeMasterPassword
      summary: Change master password
      description: Current body shape. Legacy clients send `PasswordRequestLegacy` (newMasterPasswordHash and key);
        accept both.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/PasswordRequest'
              - $ref: '#/components/schemas/PasswordRequestLegacy'
      responses:
        '200':
          description: Success, empty body.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-source-version: web-v2026.9.1
      x-source-file: libs/common/src/auth/services/master-password/master-password-api.service.implementation.ts
...
  /api/accounts/profile:
    get:
      tags:
      - Accounts
      operationId: getProfile
      summary: Get the current user's profile
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-source-version: web-v2026.9.1
      x-source-file: libs/common/src/services/api.service.ts
    put:
      tags:
      - Accounts
      operationId: updateProfile
      summary: Update profile name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-source-version: web-v2026.9.1
      x-source-file: libs/common/src/services/api.service.ts
...
  /api/accounts/revision-date:
    get:
      tags:
      - Accounts
      operationId: getRevisionDate
      summary: Get account revision date
      description: Clients compare this with their stored value to decide whether to sync.
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                type: integer
                format: int64
                description: Milliseconds since the Unix epoch of the last vault change. Returned as a bare JSON
                  number.
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-source-version: web-v2026.9.1
      x-source-file: libs/common/src/services/api.service.ts
  /api/accounts/rotate-api-key:
    post:
      tags:
      - Accounts
      operationId: rotateApiKey
      summary: Rotate the personal API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretVerificationRequest'
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-source-version: web-v2026.9.1
      x-source-file: libs/common/src/services/api.service.ts
  /api/accounts/security-stamp:
    post:
      tags:
      - Accounts
      operationId: rotateSecurityStamp
      summary: Rotate the security stamp (log out everywhere)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretVerificationRequest'
      responses:
        '200':
          description: Success, empty body.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-source-version: web-v2026.9.1
      x-source-file: libs/common/src/services/api.service.ts
...
  /api/accounts/verify-password:
    post:
      tags:
      - Accounts
      operationId: verifyPassword
      summary: Verify the master password
      description: Returns the master password policy that applies to the user, or an empty body when none applies.
        A wrong password yields 400.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretVerificationRequest'
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/MasterPasswordPolicyResponse'
                - type: 'null'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-source-version: web-v2026.9.1
      x-source-file: libs/common/src/auth/services/user-verification/user-verification-api.service.ts
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/identity/connect/token`.
    sendAccessToken:
      type: http
      scheme: bearer
      description: Send access token from the `send_access` grant (not the user token).
    hubAccessToken:
      type: apiKey
      in: query
      name: access_token
      description: User access token passed as a query parameter on the WebSocket upgrade.
  responses:
    BadRequest:
      description: Validation or business rule failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseModel'
    Unauthorized:
      description: Missing, invalid or expired access token. Clients refresh once, then sign out.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseModel'
...
  schemas:
    ApiKeyResponse:
      type: object
      properties:
        apiKey:
          type: string
        revisionDate:
          type: string
          format: date-time
      required:
      - apiKey
...
    EmailRequest:
      type: object
      description: Confirm an email change. Sent as `EmailRequest.newConstructor` in newer clients with the same
        property names.
      properties:
        masterPasswordHash:
          type:
          - string
          - 'null'
        otp:
          type:
          - string
          - 'null'
        newEmail:
          type: string
        newMasterPasswordHash:
          type: string
        token:
          type: string
        key:
          $ref: '#/components/schemas/EncString'
      required:
      - newEmail
      - newMasterPasswordHash
      - token
      - key
    EmailTokenRequest:
      type: object
      properties:
        masterPasswordHash:
          type:
          - string
          - 'null'
        otp:
          type:
          - string
          - 'null'
        authRequestAccessCode:
          type:
          - string
          - 'null'
        newEmail:
          type: string
      required:
      - newEmail
...
    KdfRequest:
      type: object
      description: Change KDF (web-v2025.8.0). Later clients perform this through the SDK and may send `authenticationData`
        and `unlockData` instead; servers should accept both.
      properties:
        masterPasswordHash:
          type: string
        newMasterPasswordHash:
          type: string
        masterPasswordHint:
          type:
          - string
          - 'null'
        key:
          $ref: '#/components/schemas/EncString'
        kdf:
          $ref: '#/components/schemas/KdfType'
        kdfIterations:
          type: integer
        kdfMemory:
          type:
          - integer
          - 'null'
        kdfParallelism:
          type:
          - integer
          - 'null'
      required:
      - masterPasswordHash
      - newMasterPasswordHash
      - key
      - kdf
      - kdfIterations
...
    KeysRequest:
      type: object
      description: Asymmetric key pair; `publicKey` is base64 SPKI DER.
      properties:
        publicKey:
          type: string
        encryptedPrivateKey:
          $ref: '#/components/schemas/EncString'
      required:
      - publicKey
      - encryptedPrivateKey
...
    PasswordRequest:
      type: object
      description: Change master password (current clients). `masterPasswordHash` is the OLD hash; the new hash
        is `authenticationData.masterPasswordAuthenticationHash`.
      properties:
        masterPasswordHash:
          type: string
        authenticationData:
          $ref: '#/components/schemas/MasterPasswordAuthenticationData'
        unlockData:
          $ref: '#/components/schemas/MasterPasswordUnlockData'
        masterPasswordHint:
          type:
          - string
          - 'null'
      required:
      - masterPasswordHash
      - authenticationData
      - unlockData
...
    SecretVerificationRequest:
      type: object
      description: 'Proof of identity. Exactly one member is normally set: the master password hash, an emailed
        OTP, or a login-with-device access code.'
      properties:
        masterPasswordHash:
          type:
          - string
          - 'null'
        otp:
          type:
          - string
          - 'null'
        authRequestAccessCode:
          type:
          - string
          - 'null'
...